LPS-74966 Move tests.
